fix to session user_id check

+++ views/update.py
@@ -12,13 +12,13 @@ def update(id:int):
 
     
     if request.method == "POST":
+        if "user_id" not in session:
+            return redirect("/")
         vocab = vocab_service.get_vocab(id)
         if session["user_id"] != vocab["user_id"]:
             flash("You cannot edit vocabs created by other users")
             return render_template("/edit.html", vocab = vocab)        
         if "delete" in request.form:
-         
-           
             result = vocab_service.delete_vocab(vocab["id"])
             if result:
                 flash(result)
